(feat) Eksponere soknad-toggles for frontend
