+++ src/backend/middlewares/escape.ts
@@ -1,7 +1,7 @@
 import { RequestHandler } from 'express';
 import xss from 'xss';
 
-import { ISøknadKontrakt } from '../frontend/typer/søknad';
+import { ISøknadKontrakt } from '../../frontend/typer/søknad';
 
 export const escapeBody: RequestHandler = async (req, _res, next) => {
     const søknad: ISøknadKontrakt = req.body;
