+++ src/backend/middlewares/feature-toggles.ts
@@ -1,33 +1,7 @@
 import { RequestHandler } from 'express';
-import { initialize, Strategy } from 'unleash-client';
 
-import { getDecorator } from './dekorator';
-
-class ByClusterStrategy extends Strategy {
-    private cluster: string = process.env.NAIS_CLUSTER_NAME ?? 'lokalutvikling';
-
-    constructor() {
-        super('byCluster');
-    }
-
-    isEnabled(parameters: Record<string, string> | undefined): boolean {
-        const clustersKommaseparert: string = parameters ? parameters['cluster'] : '';
-        const clusters = clustersKommaseparert.split(',');
-
-        return !!clusters.find(enabletForCluster => enabletForCluster === this.cluster);
-    }
-}
-
-const unleash = initialize({
-    url: 'https://unleash.nais.io/api/',
-    appName: process.env.NAIS_APP_NAME ?? 'familie-ba-soknad',
-    strategies: [new ByClusterStrategy()],
-});
-
-export const isEnabled = (feature: string): boolean => {
-    // Hvis vi bare deconstructer og eksporterer isEnabled fra unleash crasher det fordi isEnabled ikke veit hva `this` er...
-    return unleash.isEnabled(feature);
-};
+import { getDecorator } from '../routes';
+import { isEnabled } from '../utils/unleash';
 
 /**
  * Express-middleware som returnerer en feil-side hvis familie-ba-soknad.disable-soknad er skrudd på i unleash
