+++ src/backend/routes/imageprocessor.ts
@@ -2,9 +2,10 @@ import bodyParser from 'body-parser';
 import { Express, RequestHandler } from 'express';
 import sharp from 'sharp';
 
-import { logError, logWarn } from '@navikt/familie-logging';
+import { logError, logInfo, logWarn } from '@navikt/familie-logging';
 
-import { jwtValidationInterceptor } from './jwt-interceptor';
+import { basePath } from '../environment';
+import { jwtValidationInterceptor } from '../middlewares/jwt-interceptor';
 
 async function prosesser(bilde: Buffer): Promise<Buffer> {
     return sharp(bilde)
@@ -33,7 +34,9 @@ const bildeProsesseringHandler: RequestHandler = async (req, res) => {
     }
 };
 
-export const konfigurerBildeProsessering = (app: Express, path: string) => {
+export const konfigurerBildeProsessering = (app: Express): Express => {
+    const path = `${basePath}konverter`;
+    logInfo(path);
     const uploadOptions = {
         inflate: true,
         limit: '20Mb',
@@ -43,4 +46,6 @@ export const konfigurerBildeProsessering = (app: Express, path: string) => {
     app.use(path, bodyParser.raw(uploadOptions));
     app.use(path, jwtValidationInterceptor);
     app.post(path, bildeProsesseringHandler);
+
+    return app;
 };
