+++ src/backend/server.ts
@@ -1,34 +1,21 @@
-import path from 'path';
-
 import compression from 'compression';
 import cookieParser from 'cookie-parser';
 import dotenv from 'dotenv';
 import express from 'express';
-import mustacheExpress from 'mustache-express';
 
 import { logInfo } from '@navikt/familie-logging';
 
-import { indexHandler } from './dekorator';
-import environment from './environment';
-import { erklaeringInterceptor } from './erklaering-interceptor';
-import { escapeBody } from './escape';
-import { expressToggleInterceptor } from './feature-toggles';
-import { konfigurerBildeProsessering } from './imageprocessor';
-import { createApiForwardingFunction } from './proxy';
+import environment, { basePath } from './environment';
+import { expressToggleInterceptor } from './middlewares/feature-toggles';
+import { konfigurerIndex, konfigurerIndexFallback } from './routes';
+import { konfigurerApi } from './routes/api';
+import { konfigurerFeatureTogglesEndpoint } from './routes/feature-toggles';
+import { konfigurerBildeProsessering } from './routes/imageprocessor';
+import { konfigurerNais } from './routes/nais';
+import { konfigurerStatic } from './routes/static';
 
 dotenv.config();
 const app = express();
-
-const basePath = process.env.BASE_PATH ?? '/';
-const frontendMappe = path.join(process.cwd(), 'dist');
-
-// Sett opp mustache templates for index.html og disabled.html
-app.set('views', frontendMappe);
-app.set('view engine', 'mustache');
-app.engine('html', mustacheExpress());
-
-// I dev-mode vil vi ikke cache index.html, siden denne oppdateres med nye js-bundles når vi endrer ting i appen
-process.env.NODE_ENV !== 'production' && app.set('view cache', false);
 
 // Alltid bruk gzip-compression på alt vi server med express
 app.use(compression());
@@ -36,29 +23,17 @@ app.use(compression());
 // Parse cookies for bruk i dekoratør-fetch
 app.use(cookieParser());
 
-// Serve alle statiske filer utenom index.html direkte fra dist-mappen
-app.use(basePath, express.static(frontendMappe, { index: false }));
-
 // Middleware for unleash kill-switch
 app.use(expressToggleInterceptor);
 
-// Sett opp middleware for input-sanitering
-app.use(`${basePath}api/soknad`, express.json());
-app.use(`${basePath}api/soknad`, erklaeringInterceptor);
-app.use(`${basePath}api/soknad`, escapeBody);
-app.use(`${basePath}api`, createApiForwardingFunction());
+konfigurerStatic(app);
+konfigurerIndex(app);
+konfigurerNais(app);
+konfigurerApi(app);
+konfigurerBildeProsessering(app);
+konfigurerFeatureTogglesEndpoint(app);
 
-// Rendrer index.html med dekoratøren
-app.get('/', indexHandler);
-
-// Nais functions
-app.get(/^\/(internal\/)?(isAlive|isReady)\/?$/, (_req, res) => res.sendStatus(200));
-
-// Bildeprosessering
-konfigurerBildeProsessering(app, `${basePath}konverter`);
-
-// Fallback, alt vi ikke treffer med andre handlere returnerer index.html
-app.get('*', indexHandler);
+konfigurerIndexFallback(app);
 
 logInfo(`Starting server on localhost: http://localhost:${environment().port}${basePath}`);
